requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
dateparser==1.2.0
defusedxml==0.8.0rc2
Django==4.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jet-bridge-base==1.8.36
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
promise==2.3
prompt-toolkit==2.0.9
psutil==5.9.8
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
Rx==1.6.3
six==1.16.0
social-auth-core==4.5.2
SQLAlchemy==1.4.51
sqlparse==0.4.4
typing_extensions==4.9.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
